+++ src/codegen/backend_registry.cpp
@@ -5,7 +5,6 @@
 #include "unity_backend.hpp"
 #include "love2d_backend.hpp"
 #include "defold_backend.hpp"
-#include "rpc_backend.hpp"
 
 namespace lex {
 
@@ -65,10 +64,6 @@ namespace {
             
             registry.register_factory("defold", []() {
                 return std::make_unique<DefoldBackend>();
-            });
-            
-            registry.register_factory("rpc", []() {
-                return std::make_unique<RPCBackend>();
             });
         }
     };